Behold! Experimental feature!!!
Added /transcribe endpoint
Added confirm dialog before deletion, no more double click
And other fixes and improvements

+++ src/components/TrackItemEditableComponent.vue
@@ -1,5 +1,5 @@
 <template>
-  <div class="track-body">
+  <div class="track-body" :title="information">
     <div style=" flex: 1; display: flex; align-items: center;" @click="onSelectTrack">
       <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" fill="currentColor" class="bi bi-file-earmark" viewBox="0 0 16 16">
         <path d="M14 4.5V14a2 2 0 0 1-2 2H4a2 2 0 0 1-2-2V2a2 2 0 0 1 2-2h5.5L14 4.5zm-3 0A1.5 1.5 0 0 1 9.5 3V1H4a1 1 0 0 0-1 1v12a1 1 0 0 0 1 1h8a1 1 0 0 0 1-1V4.5h-2z"/>
@@ -27,12 +27,12 @@
           <path fill-rule="evenodd" d="M1 13.5A1.5 1.5 0 0 0 2.5 15h11a1.5 1.5 0 0 0 1.5-1.5v-6a.5.5 0 0 0-1 0v6a.5.5 0 0 1-.5.5h-11a.5.5 0 0 1-.5-.5v-11a.5.5 0 0 1 .5-.5H9a.5.5 0 0 0 0-1H2.5A1.5 1.5 0 0 0 1 2.5v11z"/>
         </svg>
 
-        <svg @dblclick="onRemoveTrackClick" xmlns="http://www.w3.org/2000/svg" width="24" height="24" fill="darkred" class="bi bi-trash3" viewBox="0 0 16 16">
+        <svg @click="onRemoveTrackClick" xmlns="http://www.w3.org/2000/svg" width="24" height="24" fill="darkred" class="bi bi-trash3" viewBox="0 0 16 16">
           <path d="M6.5 1h3a.5.5 0 0 1 .5.5v1H6v-1a.5.5 0 0 1 .5-.5ZM11 2.5v-1A1.5 1.5 0 0 0 9.5 0h-3A1.5 1.5 0 0 0 5 1.5v1H2.506a.58.58 0 0 0-.01 0H1.5a.5.5 0 0 0 0 1h.538l.853 10.66A2 2 0 0 0 4.885 16h6.23a2 2 0 0 0 1.994-1.84l.853-10.66h.538a.5.5 0 0 0 0-1h-.995a.59.59 0 0 0-.01 0H11Zm1.958 1-.846 10.58a1 1 0 0 1-.997.92h-6.23a1 1 0 0 1-.997-.92L3.042 3.5h9.916Zm-7.487 1a.5.5 0 0 1 .528.47l.5 8.5a.5.5 0 0 1-.998.06L5 5.03a.5.5 0 0 1 .47-.53Zm5.058 0a.5.5 0 0 1 .47.53l-.5 8.5a.5.5 0 1 1-.998-.06l.5-8.5a.5.5 0 0 1 .528-.47ZM8 4.5a.5.5 0 0 1 .5.5v8.5a.5.5 0 0 1-1 0V5a.5.5 0 0 1 .5-.5Z"/>
         </svg>
       </template>
       <template v-else>
-        <svg @dblclick="onStorageRemoveTrackClick" xmlns="http://www.w3.org/2000/svg" width="24" height="24" fill="darkred" class="bi bi-database-x" viewBox="0 0 16 16">
+        <svg @click="onStorageRemoveTrackClick" xmlns="http://www.w3.org/2000/svg" width="24" height="24" fill="darkred" class="bi bi-database-x" viewBox="0 0 16 16">
           <path d="M12.096 6.223A4.92 4.92 0 0 0 13 5.698V7c0 .289-.213.654-.753 1.007a4.493 4.493 0 0 1 1.753.25V4c0-1.007-.875-1.755-1.904-2.223C11.022 1.289 9.573 1 8 1s-3.022.289-4.096.777C2.875 2.245 2 2.993 2 4v9c0 1.007.875 1.755 1.904 2.223C4.978 15.71 6.427 16 8 16c.536 0 1.058-.034 1.555-.097a4.525 4.525 0 0 1-.813-.927C8.5 14.992 8.252 15 8 15c-1.464 0-2.766-.27-3.682-.687C3.356 13.875 3 13.373 3 13v-1.302c.271.202.58.378.904.525C4.978 12.71 6.427 13 8 13h.027a4.552 4.552 0 0 1 0-1H8c-1.464 0-2.766-.27-3.682-.687C3.356 10.875 3 10.373 3 10V8.698c.271.202.58.378.904.525C4.978 9.71 6.427 10 8 10c.262 0 .52-.008.774-.024a4.525 4.525 0 0 1 1.102-1.132C9.298 8.944 8.666 9 8 9c-1.464 0-2.766-.27-3.682-.687C3.356 7.875 3 7.373 3 7V5.698c.271.202.58.378.904.525C4.978 6.711 6.427 7 8 7s3.022-.289 4.096-.777ZM3 4c0-.374.356-.875 1.318-1.313C5.234 2.271 6.536 2 8 2s2.766.27 3.682.687C12.644 3.125 13 3.627 13 4c0 .374-.356.875-1.318 1.313C10.766 5.729 9.464 6 8 6s-2.766-.27-3.682-.687C3.356 4.875 3 4.373 3 4Z"/>
           <path d="M12.5 16a3.5 3.5 0 1 0 0-7 3.5 3.5 0 0 0 0 7Zm-.646-4.854.646.647.646-.647a.5.5 0 0 1 .708.708l-.647.646.647.646a.5.5 0 0 1-.708.708l-.646-.647-.646.647a.5.5 0 0 1-.708-.708l.647-.646-.647-.646a.5.5 0 0 1 .708-.708Z"/>
         </svg>
@@ -44,6 +44,7 @@
 <script setup>
 import { computed, inject, onMounted, ref } from 'vue';
 import { maxVisibleStringLength } from '../constants/NumericConstants.js';
+import { sysLocale, dateTimeFormatOptions } from '../constants/DateConstants.js';
 import { Status } from '../constants/StatusConstants.js';
 import { RemoveTrack } from '../functions/StorageHelper.js';
 import TrackService from '../services/TrackService.js';
@@ -76,6 +77,13 @@ const emit = defineEmits([
   'edit-track', 
 ]);
 
+const information = computed(() => {
+  const title = `${track.data.trackName} - ${track.data.artistName}`;
+  const formattedDate = new Date(track.data.dateCreated + 'Z').toLocaleString(sysLocale, dateTimeFormatOptions);
+  return `${title}
+Date uploaded: ${formattedDate}`;
+});
+
 const manipulated = computed(() => {
   let title = 'Foo';
   let artistName = 'Bar';
@@ -104,9 +112,16 @@ const onEditClick = () => {
 };
 
 const onRemoveTrackClick = async () => {
-  let response = await CaptionService.GetCaptionsByTrackId(track.data.trackId);
-  const captionId = !!response.objects ? response.objects[0].captionId : 0;
-  await CaptionService.DeleteCaptions(captionId);
+  const message = `Remove ${track.data.trackName}?`;
+  if (!confirm(message)) {
+    return;
+  }
+
+  if (track.data.hasCaptions) {
+    let response = await CaptionService.GetCaptionsByTrackId(track.data.trackId);
+    const captionId = response.objects[0].captionId;
+    await CaptionService.DeleteCaptions(captionId);
+  }
   
   await StatsService.DeleteStatsOfTrack(track.data.trackId);
   
@@ -130,6 +145,11 @@ const onRemoveTrackClick = async () => {
 };
 
 const onStorageRemoveTrackClick = async () => {
+  const message = `Remove ${track.data.trackName}?`;
+  if (!confirm(message)) {
+    return;
+  }
+
   await RemoveTrack(track.data);
   notifyRefreshFeed();
 };
@@ -137,6 +157,7 @@ const onStorageRemoveTrackClick = async () => {
 
 <style scoped>
 .track-body {
+  cursor: pointer;
   margin: 8px 4px 4px 4px;
   padding: 4px;
   display: flex;
@@ -151,7 +172,6 @@ const onStorageRemoveTrackClick = async () => {
 }
 
 .track-body:hover {
-  cursor: pointer;
   color: white;
   border-style: dotted;
   border-width: 2px;
